requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.28.0
